requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
langgraph>=0.2.0
langchain-core>=0.2.0
Pillow>=10.0.0
numpy>=1.24.0
plotly>=5.18.0
python-dotenv>=1.0.0